requirements.txt
Flask==2.3.2
numpy>=1.25.0
Pillow==9.5.0
tensorflow==2.18.0
scikit-learn>=1.3.1
